Fix test imports
